+++ src/components/Loading.tsx
@@ -1,0 +1,14 @@
+export default function LoadingPage() {
+  return (
+    <div className="flex items-center justify-center min-h-screen bg-gradient-to-br from-amber-50 to-orange-50">
+      <div className="text-center">
+        <div className="inline-block h-16 w-16 animate-spin rounded-full border-4 border-solid border-orange-400 border-r-transparent align-[-0.125em] motion-reduce:animate-[spin_1.5s_linear_infinite]" role="status">
+          <span className="!absolute !-m-px !h-px !w-px !overflow-hidden !whitespace-nowrap !border-0 !p-0 ![clip:rect(0,0,0,0)]">
+            Loading...
+          </span>
+        </div>
+        <p className="mt-6 text-xl font-semibold text-orange-900">Loading Beverage Society...</p>
+      </div>
+    </div>
+  );
+}
